refactor(claude): isolate claude-swap adapter responsibilities
